requirements.txt
fastapi[all]
openai
canvasapi
pypdf
python-docx